some changes

+++ admin/rental_overview.php
@@ -60,7 +60,7 @@ require '../process/admin_secure.php';
               <td><?php echo $date[0] . "-" . $date[1] . "-" . ($date[2] + $orderlist['lob']); ?></td>
               <td>
                 <a href="rental_view.php?id=<?php echo $orderlist['cid']; ?>"><button class='button button-green'>View more</button></a>
-                <a href="../process/delete_order.php?id=<?php echo $orderlist['cid']; ?>"><button class='button button-red'>Delete</button></a>
+                <a href="../process/delete_order.php?id=<?php echo $orderlist['cid']; ?>"><button class='button button-red' onclick="return confirm('Are you sure');">Delete</button></a>
               </td>
             </tr>
           <?php
@@ -70,12 +70,11 @@ require '../process/admin_secure.php';
       </section>
     </main>
   </div>
-  <footer class="footer admin-footer">
-    <div>
-      <p style="color: white">&copy;2023 Ezy Rental .All rights reserved.</p>
+  <!-- <======== FOOTER ========> -->
+  <?php
+  include '../include/admin_footer.php';
+  ?>
 
-    </div>
-  </footer>
 </body>
 
 </html>
